backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
scikit-learn==1.3.0
openai==1.3.0
gunicorn==21.2.0
eventlet==0.33.3
SQLAlchemy==2.0.21
Werkzeug==2.3.7 